requirements.txt
pandas
numpy
yfinance
scipy
python-dotenv
requests
openai
pyarrow
fastparquet